Restructured the code for clarity
